+++ packages/stream_chat/lib/src/repository/devices_repository.dart
@@ -2,7 +2,7 @@ import 'package:stream_core/stream_core.dart' show PatternMatching, Result;
 
 import '../../open_api/api.dart' as api;
 import '../core/models/device.dart';
-import '../core/models/responses/list_devices_response.dart';
+import '../core/models/list_devices_response.dart';
 import 'mapper/devices_mapper.dart';
 
 /// Repository dedicated to device operations.